Add files via upload
Updated the program so that it deals with information sets, instead of single inputs. The net loops through the input, one set at a time, or through all sets per epoch to train, depending on which function the user calls. These training methods are terminated either by some error threshold or by iteration.